pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "systemair-api-promises"
version = "0.1.0"
description = "Python library for communicating with and controlling Systemair ventilation units"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Henning Berge", email = "henning.ber@gmail.com"}
]
maintainers = [
    {name = "Henning Berge", email = "henning.ber@gmail.com"}
]
keywords = ["systemair", "ventilation", "hvac", "home-automation", "api"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Home Automation",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Operating System :: OS Independent",
]
requires-python = ">=3.7"
dependencies = [
    "requests>=2.28.0",
    "websocket-client>=1.3.1",
    "beautifulsoup4>=4.10.0",
    "python-dotenv>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=3.0",
    "pytest-mock>=3.6",
    "requests-mock>=1.9",
    "black>=22.0",
    "isort>=5.10",
    "flake8>=4.0",
    "mypy>=0.950",
    "sphinx>=4.0",
    "sphinx-rtd-theme>=1.0",
    "sphinx-autodoc-typehints>=1.12",
    "pre-commit>=2.15",
    "build>=0.8",
    "twine>=4.0",
]

[project.urls]
"Homepage" = "https://github.com/promises/SystemAIR-API"
"Documentation" = "https://github.com/promises/SystemAIR-API#readme"
"Repository" = "https://github.com/promises/SystemAIR-API"
"Bug Tracker" = "https://github.com/promises/SystemAIR-API/issues"
"Changelog" = "https://github.com/promises/SystemAIR-API/blob/main/CHANGELOG.md"

[project.scripts]
systemair-api = "systemair_api.__main__:main"

[tool.black]
line-length = 88
target-version = ["py39", "py310"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_any_unimported = false
no_implicit_optional = true
strict_optional = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"